// apps/tmac-apps/src/types/components.types.ts
export type BackgroundImage = {
  backgroundImageUrl: string;
  backgroundImageSize: string;
};

export type BaseComponentStyles = {
  color?: string;
  backgroundColor?: string;
  backgroundImage?: BackgroundImage;
  borderTop?: string;
  borderLeft?: string;
  borderRight?: string;
  borderBottom?: string;
  outline?: string;
  fontWeight?: string;
  boxShadow?: string;
  borderRadius?: string;
  border?: string;
  width?: string;
};